[WIP] WP

+++ packages/orchestrator/internal/sandbox/uffd/userfaultfd/mocks_test.go
@@ -14,11 +14,11 @@ const pagesInTestData = 32
 
 type mockMappings struct {
 	start    uintptr
-	size     int64
-	pagesize int64
+	size     uint64
+	pagesize uint64
 }
 
-func newMockMappings(start uintptr, size, pagesize int64) *mockMappings {
+func newMockMappings(start uintptr, size, pagesize uint64) *mockMappings {
 	return &mockMappings{
 		start:    start,
 		size:     size,
@@ -26,11 +26,11 @@ func newMockMappings(start uintptr, size, pagesize int64) *mockMappings {
 	}
 }
 
-func (m *mockMappings) GetRange(addr uintptr) (int64, int64, error) {
+func (m *mockMappings) GetRange(addr uintptr) (uint64, uint64, error) {
 	offset := addr - m.start
 	pagesize := m.pagesize
 
-	return int64(offset), pagesize, nil
+	return uint64(offset), pagesize, nil
 }
 
 type mockSlicer struct {
@@ -45,15 +45,15 @@ func (s *mockSlicer) Slice(offset, size int64) ([]byte, error) {
 	return s.content[offset : offset+size], nil
 }
 
-func newMock4KPageMmap(size int64) ([]byte, uintptr) {
+func newMock4KPageMmap(size uint64) ([]byte, uintptr) {
 	return newMockMmap(size, header.PageSize, 0)
 }
 
-func newMock2MPageMmap(size int64) ([]byte, uintptr) {
+func newMock2MPageMmap(size uint64) ([]byte, uintptr) {
 	return newMockMmap(size, header.HugepageSize, unix.MAP_HUGETLB|unix.MAP_HUGE_2MB)
 }
 
-func newMockMmap(size, pagesize int64, flags int) ([]byte, uintptr) {
+func newMockMmap(size, pagesize uint64, flags int) ([]byte, uintptr) {
 	l := int(math.Ceil(float64(size)/float64(pagesize)) * float64(pagesize))
 	b, err := syscall.Mmap(
 		-1,
@@ -69,16 +69,16 @@ func newMockMmap(size, pagesize int64, flags int) ([]byte, uintptr) {
 	return b, uintptr(unsafe.Pointer(&b[0]))
 }
 
-func repeatToSize(src []byte, size int64) []byte {
+func repeatToSize(src []byte, size uint64) []byte {
 	if len(src) == 0 || size <= 0 {
 		return nil
 	}
 
 	dst := make([]byte, size)
-	for i := 0; i < int(size); i += len(src) {
-		end := i + len(src)
-		if end > int(size) {
-			end = int(size)
+	for i := uint64(0); i < size; i += uint64(len(src)) {
+		end := i + uint64(len(src))
+		if end > size {
+			end = size
 		}
 		copy(dst[i:end], src[:end-i])
 	}
@@ -86,7 +86,7 @@ func repeatToSize(src []byte, size int64) []byte {
 	return dst
 }
 
-func prepareTestData(pagesize int64) (data *mockSlicer, size int64) {
+func prepareTestData(pagesize uint64) (data *mockSlicer, size uint64) {
 	size = pagesize * pagesInTestData
 
 	data = newMockSlicer(
